pyproject.toml
[tool.poetry]
name = "argonautai"
version = "0.1.0"
description = "An Argo workflows deploy utility tool for me."
authors = ["nagomiso <daiseigi.giants+github@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.3"
"ruamel.yaml" = "^0.17.16"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^21.9b0"
isort = "^5.9.3"
mypy = "^0.910"
pytest = "^6.2.5"
pre-commit = "^2.15.0"
pylint = "^2.11.1"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.2"
testpaths = [
    "tests",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
